refactor(compiler): move collectLoopChildConditionals to break require() cycle
Move `collectLoopChildConditionals` from reactivity.ts to collect-elements.ts
so it sits next to `collectBranchConditionals` and `summarizeBranch` — all
three build per-branch metadata (HTML templates, inner-loop metadata, child
components, nested conditionals) rather than classifying reactive expressions.

Eliminates the lazy `require('./html-template')` / `require('./collect-elements')`
inside the function body, which only existed to work around the
reactivity.ts ↔ collect-elements.ts import cycle.

Pure refactor: baseline test results unchanged (580 pass / 11 fail /
7 errors, same as parent PR).
